fix(hooks): end-to-end minting and burning
