+++ src/hooks/useInterval.ts
@@ -1,0 +1,21 @@
+import { useEffect, useRef } from "react";
+
+export function useInterval(callback: any, delay: number) {
+  const savedCallback = useRef(() => ({}));
+
+  // Remember the latest callback.
+  useEffect(() => {
+    savedCallback.current = callback;
+  }, [callback]);
+
+  // Set up the interval.
+  useEffect(() => {
+    function tick() {
+      savedCallback.current();
+    }
+    if (delay !== null) {
+      const id = setInterval(tick, delay);
+      return () => clearInterval(id);
+    }
+  }, [delay, savedCallback]);
+}
